JVMCBC-993: Optimize early discard of events which are not going to be logged
Motivation
----------
During profiling we discovered that the logging handler would do all kinds of
work (including preparing the log message description) and send it down into
the logger, even though the event is below the set logger verbosity level.

This causes unnecessary work and garbage collection.

Modifications
-------------
The logger is now created / accessed much earlier and simple verbosity
checking is put in place so that the logging event consumer can bail
out as early as possible on an event which has no chance of being
logged anyways.

Result
------
Less work to do and less garbage to collect.

Change-Id: Iac0d1ded2f9e2eda064f7939a19ff4f02b6f3781
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/157304
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>