JVMCBC-996: Throw FeatureNotAvailableException if scope level queries are not available
For both query and analytics, it's important to return / throw a feature not found
error if scope level queries are not supported. The query service returns a specific
error which we can latch onto (so there is no preflight check needed), but analytics
does not follow the same pattern.

So for analytics we need to do a little more work upfront and compare the request
with the config available to see if collections are part of the bucket capabilities.
If they are not, we fail fast and not even send the request to the server.

Change-Id: Ib0ba4c655a2d6d6a6a8bb03a1591ae06aae77564
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/157415
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/158044
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>