JVMCBC-998: Don't set tracing spans if not needed/consumed
Both the Noop and the ThresholdLogging Tracers do not actually
look at the spans (the threshold tracer gets fed with the
request context directly for performance reasons), so in those
cases as an optimization we can avoid doing the span attribute
work in the first place.

Note that not all attribute setters are optimized "away", only
those which are deemed to be performance critical to strike
a balance between code clutter and perf benefits.

Change-Id: I7dfd7b4cddfd20668b7be1a6ea148767a4b275c7
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/157433
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>