JVMCBC-999: Properly map server query timeout while streaming.
Motivation
----------
It has been discovered that when the server starts sending rows
but times out in the middle of sending, the query fails with an
error which is not mapped properly in the SDK (to a timeout).

Usually the client always times out more quickly than the server
does, but in this case the timeout is already cancelled on the
client side when the streaming starts.

Modifications
-------------
This changeset properly maps a timeout during streaming received
from the server to a client side unambiguous timeout.

Change-Id: I7210e807327f94c63118105d7b7202ce7cbde864
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/157762
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>