JVMCBC-990: Gracefully handle cluster-level WaitUntilReady pre 6.5
Motivation
----------
It has been discovered that when WaitUntilReady at the cluster level is
performed against clusters which do not have support for cluster level
configs (so all clusters pre 6.5), if one of the seed nodes is unresponsive
the app is not able to make progress.

Note that there is no "perfect" solution here, since on the one hand
we can error and tell the app user to not perform that op against older
clusters, but on the other hand there is no harm in moving on, since
the consecutive bucket open will likely succeed and the app can make
progress this way.

As a result, this what we do here.

Modifications
-------------
This changeset removes an outdated early check which raised an error
before, but was very much dependent on timing and not helpful at all.

Instead, it now gracefully checks down in the process once the bootstrap
attempt is completed if we are in a "gcccp" situation and if we could not
resolve a config. In this case, we gracefully back out and do not
throw an error (and more imporantly try ping etc.), but print a INFO
level message which explains what's going on and what the user, ideally,
should do instead.

The WaitUntilReadyContext is also present and has been extended to
account for this situation.

Result
------
Graceful handling of cluster-level WaitUntilReady against pre 6.5
clusters.

Change-Id: I6a71c22a528ae602cdd4ed1f9fb6e9b6d752ad6a
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/158064
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>