JVMCBC-1002: Revert ConsoleLogger being the default.
This changeset reverts a change that made it into the 3.2.0 SDK
line that, if no logger otherwise is present, that the consoler
logger should be the default.

After this change, the JUL logging is used as the default if none
other present, and users can switch to console logging through
config options.

Change-Id: Ie6844aa1bf3e42e0ae7ca67ad878681b29c83576
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/158470
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>