JVMCBC-1006: ErrorContext must be included in message
This change makes sure that the error context is included
in the message of the CouchbaseException, not just in
toString. This is important for setups where the message
and the stack is printed individually and not through
toString. In those cases the error context is not printing.

Also to prevent a user from accidentially overriding the
message and prevent the context from printing, it is
marked as final in the CouchbaseException itself.

Change-Id: I703c7a299f7b1b07df4ad3d7e3268b22edc4fd45
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/159357
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>