JCBC-1868 Make JsonValueModule compatible with Jackson < 2.8
Motivation
----------
If a user added their own ancient Jackson 2 to the class path,
calls to contentAsObject() / contentAsArray() would fail
because JsonValueModule used a method that's only present
in Jackson 2.8 and later.

Modifications
-------------
Make JsonValueModule compatible with any version of Jackson 2
by calling JsonParser.getCurrentToken() instead of
JsonParser.currentToken().

The change is only strictly necessary in JsonValueModule,
but apply the change to the Repackaged flavor as well since
we want to keep the two flavors in sync.

Change-Id: I68bd605f56d4787b0d5e9bd0260147ffcb50b51c
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/159928
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>