JVMCBC-1013 Move N1QL prepared statement cache into core-io
Motivation
----------
Reuse prepared statement code in different JVM client implementations.

Modifications
-------------
Split the prepared statement code into two separate strategy classes.
Start with the legacy strategy, and upgrade to the enhanced strategy
when the cluster config says it's supported.

Stop listening to the clustermap configs after the enhanced strategy
is enabled, since this is a one-way gate.

To avoid using JsonObject and QueryOptions (which are specific
to the Java SDK), add new methods QueryRequest.toPrepareRequest()
and toExecuteRequest() which create specialized copies of the request
without needing to re-inject the fields from QueryOptions.

Mark PREPARE statements as not idempotent if auto-execution is used
and the original statement is not idempotent.

Change-Id: I1cf8eaaef841fbe268f31d713d745ceebc77f0dd
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/160363
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>