JVMCBC-1017: Threshold Logging Tracing not working due to context not being set
This changeset fixes a regression introduced in 2.2.0 with change number
2aa30aabac5618be018ceb650fc264ae49b578b2 where the RequestContext is not set
at all on the threshold spans which prevents the threshold logging tracer
from working properly. It does not affect the meter functionality and all
the other tracing modules (otracing and otel) still work fine.

A integration test has been finally added that catches this so it doesn't
happen again.

Change-Id: Ibeab26a8c24e7cd0d93f2075bd34ff7eb0811017
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/161681
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>