JVMCBC-1022: Batch-Log overflow messages in the DefaultEventBus
This changeset modifies the DefaultEventBus so that events which
are overflowing are not directly logged to stderr but rather
"batched up" and logged at a given interval.

Note that of course now there is a "loss of precision", since
not all dropped events are logged - but we preserve one event
per type and print it as part of the message.

Change-Id: Ida6abe7c7757338c0126c1ea5b9d1b81de91e373
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/162879
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>