JVMCBC-1032: Add support for Rate/Quota Limits
Motivation
----------
This change is needed to provide a better user experience
for an upcoming feature (in 7.1): rate limits and quota
limits. Ops would fail with older clients too, but in this
changeset we make sure that they are surfaced correctly,
with proper descriptions and context.

Modifications
-------------
Each participating service (at the moment kv, search, query
and the cluster manager) handles the corresponding error codes
or error strings and turns them into one of the two new exceptions:
rate limit failure or quota limit failure - depending on the type
of limiting.

These errors are never retried at the moment. A big bunch of integration
tests has been added to verify the basic functionality, although not all
the different cases (there are many combinations) are fully tested.

Result
------
Support for rate and quota limits implemented.

Change-Id: Iebbe66617033772cc81ee5db609d27bf26efca32
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/165001
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>