JVMCBC-1041: Fix manager error propagation.
This changeset fixes a regression in the currently unreleased
version that has been introduced when adding the rate limiting
support at the manager level.

The change made it return a generic CouchbaseException as a
catch-all vs. the HttpStatusCodeException that the higher level
code was expecting.

This changeset reestablishes that response, which will in turn
make the integration tests pass again.

Change-Id: Ib3290722e48cc5a2038dd8b23726c27968705fd1
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/165924
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>