JVMCBC-1045: Implement basic state Watchdog in Core
Motivation
----------
A watchdog is introduced into the Core which should make sure
that the system looks "sane", and if not try to take corrective action.

The first watchdog check introduced by this change is checking the
number of managed nodes vs. the number of nodes in the configuration.
If something looks off (they differ in the number), a reconfiguration
is triggered. This is run every 5s.

Change-Id: Ieb58e01730c427d0603ba6962c1dcc9db276db53
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/167109
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>