JCBC-1896 JsonValueSerializerWrapper does not support decoding via TypeRef
Motivation
----------
When a user supplies a customer JsonSerializer, we wrap it in a
JsonValueSerializerWrapper so it can always handle JsonObject and JsonArray.

The issue is that JsonValueSerializerWrapper fails to delegate to
the target serializer when deserializing using a TypeRef.

Modifications
-------------
Delegate to the target serializer instead of using the default method
which throws an exception.

Change-Id: Ia2980810929b0be49843bf3b605e65a95c2cf7c2
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/167523
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/171163
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>