JCBC-1902, SCBC-342: Send configured user timeout to Search.
Right now the server will apply a 10s default timeout if not
provided as part of the parameters. Java did not apply a timeout
at all on the server side, scala only did when the serverSideTimeout
is manually specified.

This changeset brings the behavior in line with other timeouts
that we send to the server, i.e. query. If the user does not
override the timeout the default configured timeout will be sent.

Note that for scala, since serverSideTimeout is also present there,
it will take precedence if configured for backwards compatibility
reasons.

Change-Id: I37c9ee7e6348e648d925493c1d301a6550752c8e
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/167746
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>