JVMCBC-1053 Throw FeatureNotAvailableException when query preserve expiry is not supported
Motivation
----------
The current CouchbaseException has a generic message of
"Unknown query error", so a user has to look in the error context
to see what's going on... and even then, there's no indication
the query failed because the server version is too low.

Modifications
-------------
Throw FeatureNotAvailableException with message about required
Couchbase Server version.

Change-Id: I82cfc5736ef53680f0016bf5b8801c170532b023
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/168860
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>