JVMCBC-1059: Improved Query Error Handling
This changeset implements support for the newly added
"reason" section in the query engine with server 7.1
or later.

This changeset consists of the following parts:

 - The ErrorCodeAndMessage class is extended to also hold
   the retry and reason fields.
 - When parsing a response, those fields are checked for
   and used if present.
 - The retry flag is now considered as part of the "send
   into the retry orchestrator" logic
 - New error codes are considered based on the document,
   for all 3 unit tests are added with payloads from the
   server.
 - The error context contains the additional information
   verbatim.

As a result, with 7.0 you will see on double insert with
the same key a DmlFailureException while on 7.1 you'll
see a DocumentExistsException. The reason also shows
as part of the error context.

Change-Id: I3059fcdb043f4f433dc0153b3fe3399c6a58dcd6
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/168919
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>