SDKQE-2555: Add integration test option to run all tests with TLS enabled
Motivations:
It is useful to have the ability to run all the Java and Scala integration tests with TLS Enabled. With the
introduction of the multi-certs testing, all integration testing was using TLS by default which was not intended,
so this patch reverts the integration tests to only run with TLS when the option is enabled, or a multi-certificate
file is provided.

Modifications:
 - Added unmanaged cluster integration test option "cluster.unmanaged.runWithTLS" enabled/disabled TLS for the tests
 - Providing the multi-certificate integration test option will always run with TLS enabled
 - Fixed an Scala integration test environment bug that was making the tests always use an empty ClusterEnvironment

Change-Id: I99b37767f8087e451b96267bec86ea2fff4cdea2
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/169356
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>