JVMCBC-1064 Allow custom RetryStrategy to specify exception type
Motivation
----------
User wants to write a custom RetryStrategy that fails fast when
a document is locked, and wants the operation to fail with
a specific exception.

Modifications
-------------
Add an overload for RetryAction.noRetry that takes an exception
translator that converts an exception into the desired type.

Pass the exception translator down into BaseRequest.cancel,
and apply it to the default exception created there.

Change-Id: Ic4501aaa1ce30b2601e42f755dbe7b77fe3a999c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/170059
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>