JCBC-1917 Simplify ClusterEnvironment configuration
Motivation
----------
If all you want to do is tweak a few config options, it's
a hassle to create a ClusterEnvironment and bear the burden
of shutting it down.

Modifications
-------------
Add ClusterOptions.environment(Consumer<ClusterEnvironment.Builder>).
When you call this method, the cluster creates and manages the environment,
giving your callback a chance to custominze the builder.

Update the integration tests to use the new method. (There was at least
one case where we were neglecting to shut down the environment.)

Add an integration test specifically for sharing a ClusterEnvironment.

Change-Id: I1f6fbffb5a6bb92384699c8db0ef79a7c7bee033
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/170140
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>