KCBC-62 Temporarily remove OptIn requirements for volatile and uncommitted API
Motivation
----------
Kotlin's opt-in system has not yet been stabilized in Kotlin 1.6.
Rather than risk incompatibility with the future stable implementation,
we're going to hold off on adding opt-in requirements.

Modifications
-------------
Features annotated with @VolatileCouchbaseApi and @UncommittedCouchbaseApi
no longer require opt-in.

Note that they will likely require opt-in in the future,
once Kotlin's opt-in feature has stabilized.

Change-Id: Ia7c5014d2fbb96bb5e53eae5b704cfa449179b91
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/170986
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>