JVMCBC-1070 Bundle Capella CA certificates with the SDK JCBC-1923 Warn when ignoring connection string parameters or scheme
Motivation
----------
Streamline configuring the SDK to connect to a Capella cluster,
so users don't need to download the CA certificate
from the Capella website.

Modifications
-------------
Bundle the Capella CA certificate with the SDK. If the user
does not configure a different trust source, trust the
Capella CA.

Retain the existing fail-fast behavior for non-Capella clusters:
Throw an exception if the user doesn't specify a trust source
when connecting to a non-Capella cluster using TLS.

If the user tries connecting to Capella without TLS,
log a warning and let the attempt continue.

Log warnings if bits of the connection string are ignored
because they are incompatible with a shared cluster environment.

Since the Capella detection code depends on having a connection
string, add a helper method for turning a set of SeedNodes
into a synthetic connection string to handle the case where
the user connects using SeedNodes.

Change-Id: I517fa083d3859a321fb37d3ab972110abbc4bc7c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172155
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>