JCBC-1929: Integrate transactions library into SDK (java-client changes)
(This is the first part of a series of changes, it will
not compile standalone.)

This work begins integrating the transactions library
directly into the Java SDK.  This first set of changes
are the ones that either touch existing SDK files or are
the parts of the interface the user will directly interact
with.  This is the vast majority of the changes in java-client,
the remaining code is in core-io and will be presented in
later changesets.

A key objective is to permit future Kotlin and Scala clients
atop a shared core, so the java-client side is kept
minimal.

Change-Id: I73edcc220ee5cb663d9cefd44b3f9be6adbb7bb4
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172347
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>