JCBC-1929: Integrate transactions library into SDK (core-io changes)
This work continues integrating the transactions library
directly into the Java SDK.  This is the core-io portion,
designed to be shared by java-client and in future the Kotlin
and Scala clients.

There are parts of this code, particularly CoreTransactionAttemptContext,
that need a huge refactor.  That is planned but probably won't be done
as part of this changeset, which is aiming for a minimally invasive
port of the existing transactions library under a tight timescale.

Change-Id: I60ca3159b765c3dbe3b496f51e5df5c274dab4fe
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172441
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>