JVMCBC-1076 Deprecate Event.createdAt() in favor of a version that returns an Instant
Motivation
----------
Event.createdAt() returns the value of System.nanoTime()
captured when the event was created. This great if you want
to answer the question, "How many nanoseconds have elapsed
since this event occurred?" but makes it inconvenient to
get the event's timestamp.

Modifications
-------------
Add a new Event.created() method that returns an Instant
representing the event's creation timestamp.

Deprecate Event.createdAt(), since nanoTime confuses everyone,
and sub-millisecond resolution is not critical for the timestamps.

Change-Id: Id58c23b96b316b831d88104e108fbaa5a6010aa2
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172674
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>