JCBC-1929: Integrate transactions library into SDK (java-fit-performer)
An important part of integrating the transactions library
into the SDK is to provide testing.

Transactions testing is done by FIT, which is based on a
driver-performer architecture where the tests are written in
Java in a driver, which drive per-language performers to
run transactions of various colours.

Adding the performer directly into the monorepo in this way
provides good development workflow, as changes can be tested
simply by restarting the performer.

The module is not added to the main top-level pom.xml at this
time, as it has a dependency on a GRPC library that must
be built separately.

Change-Id: Ibe562796a81781024b2aa3ab22e60f94b2c317f3
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>