tests: try to make mockito not whine about cleansIdleConnections.
This test completes most of the time, but it fails here and there.
Since we are not actually stubbing a void method with a return value,
the internet suggests that this could be due to a threading/race
issue.

So this changeset tries a couple things to improve the situation,
time and CI will tell if it helps.

Change-Id: I3f2385c8e6df4bbac692c4b1637e608af1759d4e
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172903
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>