JCBC-1929: Integrate transactions library into SDK (core-io followup) TXNJ-472: Rework for ExtSingleQuery
We are tweaking the behaviour of single query transactions
(e.g. those initiated from cluster.query()) so that
most errors, such as ParsingFailedException, are raised
directly rather than being wrapped in a TransactionFailed.
The intent is that the errors are easier for the user
to deal with, e.g. if they already have handling for a
particular query error.

Since the SDK integration is not yet published, doing this
work under the same JCBC-1929 ticket.

Change-Id: I3529554d73bdecf58f438c0ceb24bec540381f1b
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/172961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>