TXNJ-266: StagedMutations do not need to store TransactionGetResult
(Doing this couchbase-jvm-clients work under the existing TXNJ
rather than a JCBC, since the code is not yet published.)

Each StagedMutation currently stores a TransactionGetResult.
This unnecessarily stores the staged content twice - once in
the TransactionGetResult, once in byte[] content.

This is a also pre-requisite for future enhancement TXNJ-357,
where the content will not always be stored in StagedMutation.

Change-Id: I4e2015e0861d68da1082ec1f95b01e7e90678205
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173236
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>