JVMCBC-1083: Also complete header when metrics show up.
This change is needed because when we hit a different analytics
endpoint, the response doesn't completely look like what we are
used to. This causes an issue when no rows are returned, since no
status is present and then we'll never complete the header and
time out.

This change also completes the header when metrics show up, which
are also reported in the alternative http endpoint.

Change-Id: I7c0ece0daae94f92b4ff34833168c085fb2e4ad6
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173318
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>