Transactions: Adding ExtMobileInterop support
(No ticket as the transactions logic is not yet published)

This complex piece of work is specified in a
separate technical document.

The short version is that it reworks
handling of CAS mismatches at all stages (rollback,
commit, and staging), often allowing the operation
to be retried rather than the entire transaction
needing to be.

In particular, and the initial driver for this change,
is that it improves compatibility with
some niche situations involving Sync Gateway.

The new functionality is tested extensively by FIT.

Change-Id: Ib1ade6b04abb6016f0cb0405cbe1ba8cfd57e2ce
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173496
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>