JVMCBC-1085: Store last connect attempt failure and expose in diagnostics.
This changeset keeps track of the last failure (if any) that happened
during the connect attempts. It is then made part of the EndpointDiagnostics
so it can be either picked up in the logs or programmatically.

This functionality is going to be used in higher level libraries to fine-tune
auth error behavior.

Change-Id: Id5c5a10f9c734fe885e887d33f85749482473325
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173557
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>