JCBC-1934 Incorrect GetAllIndexes response on default collection
Motivation
----------
When QueryIndexManager.getAllIndexes was limited to the default scope,
(or the default scope and default collection), indexes on the default
collection were not included in the results.

Modifications
-------------
When the results should include the default collection, add the
special bit of N1QL that matches indexes on the default collection.

Add a CoreQueryIndexManager with helper methods for making the
N1QL statement and parameters.

Change-Id: I31016e23f1a8c9ae12deeaa86be5d91d2147f26b
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173590
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>