KCBC-74 Add Expiry.Unknown and make GetResult.expiry non-null
Motivation
----------
Working with nullable properties is painful.

Also, it was never completely obvious that a null Expiry
meant the expiry was unknown -- it was too easy to assume
it meant there was no expiry.

Modifications
-------------
Add Expiry.Unknown, and use it instead of null to represent
unknown expiry.

Throw IllegalArgumentException if the user tries to use
Expiry.Unknown to set a document's expiry.

This is a BREAKING API CHANGE.

Change-Id: Ib6ac2f4d3011f1249c6ed569ec446f647c53edbe
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173659
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>