Jenkinsfile - remove stashing
Seeing issues on CI where source is being used (based
on stacktraces and log output) that doesn't correspond
to the source that is supposed to be checked out.
But, only on some files - others are up-to-date with
recent changes.

The cause of this isn't clear.  As an experiment,
removing all stashing, and replacing cleanWs() with
deleteDir().  The documentation for cleanWs() is
unclear on whether it's an imperative instruction
to clear the workspace now, or a declarative instruction
to remove the workspace after the build.  The
semantics for deleteDir() are clear.

Change-Id: I094e3434586cee51bfbf5f8c190d4ae0f7f786e8
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/173684
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>