Jenkinsfile - improving agent use
Most of the stages are hardcoded to be on a specific agent
when it looks like there's nothing preventing them using
any of the sdkqe, which will better distribute the work.

Also using beforeAgent, which will evaluate the `when` before
an agent is assigned to the stage.  This avoids needlessly
using agents.

Removing the package stage, which is not working and not used
by anything.

Change-Id: I156d5306699eff001589cad0a63412db2ef73fca
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/174050
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>