JCBC-1955: Running too many concurrency transactions hangs
When X transactions are run concurrently with the blocking API
(where 50 < X <= 99), they hang.  This is because they're all
running the user's lambda on the default transactions scheduler
which has a small thread pool as it's not meant to be used for
anything that will block for long.  They should be running on
schedulerBlocking, which is effectively unlimited.

So the main change is in ReactiveTransactions, which now runs the
user's lambda on that blocking scheduler, in the blocking API.

This also instigated a hard look at the transactions scheduler
usage, which prompted some related changes.

Have now simplified the transaction schedulers from 3 to 2.  I
no longer see the benefit of a separate 'default' transaction
scheduler.  We will have one scheduler for everything that
happens in 'user space', and is an unlimited thread pool.
We'll run the user's lambda on this
in blocking mode (which means one thread per lambda).  And
all operations (KV & query) will switch to this after returning
from the SDK, so in the reactive API we are again on this
unlimited pool when we return to user space.  This avoids problems
if the user accidentally does a blocking op on that - it won't
tie up a very limited SDK thread.

TransactionAttemptContext (the blocking API): removing the
publishOn statements.  There's no need for these, as all
operations have already been publishOn'ed to schedulerLambda.
(This code is likely a holdover from before that was done
thoroughly.)

Change-Id: I9f2536d570d94cb8a2484e7f860186018cfae5c2
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/175186
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>