JVMCBC-1103: Only fetch one config per poll interval in the refresher.
Right now we are fetching up to 3 from 3 different nodes. This allows us to more quickly
react to failure, but comes with some overhead. Since the poll interval is already low
(at 2.5s), checking one node in a round-robin fashion is enough.

This also aligns it with the other SDKs in terms of behavior.

Change-Id: I8eccfb1ffcc8fec9387ba88218f86a0dd0d4b160
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/176101
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>