JCBC-1975: Transactions - support ExtReplaceBodyWithXattr
This extension leverages a new feature added in memcached in 7.0,
and then fixed in 7.1, which adds a Sub-Document spec that copies
the contents of an xattr field into the same document's body.

This was added for transactions, and provides an improved solution for
commiting documents that replaces ExtTimeOptUnstaging (which holds onto
the staged data in memory) and ExtMemoryOptUnstaging (which instead
rereads the document prior to commit).

The change involves checking if the cluster supports this feature
up-front, so we know if we need to fallback to ExtTimeOptUnstaging.
And if it does support it, we use that Sub-Document spec at commit
time.

Change-Id: I02721093ce0f874e4df7becc21f83166e6e6e2b3
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/176520
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>