JVMCBC-1112: Replace Chunk Trailer MonoProcessor with Sinks.One
This is a first commit of a series in an effort to replace the
deprecated Processor APIs with the new set of supported Sinks
APIs in Reactor.

In this commit, the chunk trailer (for i.e. queries) is replaced
with a Sinks.One. Note that there is no need to call onComplete
anymore, this is done by tryEmitValue automatically.

Change-Id: I5503b97bcaea1a72eb2a550168859346c7dfab7b
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/177457
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>