JVMCBC-1115: Allow for configurations with only TLS ports.
Motivation
----------
At the moment the SDK works under the assumption that a config returned always
contains non-tls ports, even if tls is mandatory. In future server deployments,
it will be possible that ONLY tls ports are provided, so the SDK needs to be
prepared for it.

Modifications
-------------
This changeset modifies the config loading code to take the fact into account
that the "directServices" might be null / not available. In these cases a fallback
to the ssl services is provided.

The code which aggregates the list of KV nodes for the partitions is changed slightly:
Instead of also checking the port in addition to the hostname, just the hostname is checked
and to avoid adding it more than once the code checks if it is already present. Since
we first make sure that the KV service is actually enabled on that node, the list
should end up identical as before and work with the new TLS only scheme were no "port"
is provided anyways.

Result
------
Work against future configs which only have TLS enabled.

Change-Id: Ia300e87c51b42226649c7c2e32be09097f279dcd
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/177653
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>