JVMCBC-1112: Fix race condition regression.
The previous changesets did not account for a race condition
of unserialized access in the Sinks which can lead to dropped events.
Based on discussion with the reactor team and best practices, all
emit events now use the busy looping optimistic conflict retry
strategy so that the emits eventually do happen.

Note that in specific critical paths external synchronization
with better error reporting is put in place (i.e. config pushing
where "bail out" is not an option).

Change-Id: I44467c49f6c8bf8b77682e1d4685d75b54e26049
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/177714
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>