JVMCBC-1119: Do not wait for num (kv) "nodesExt" == num "nodes" on Bucket waitUntilReady
Motivation
----------
When attempting to improve the eventually consistent aspects of a Couchbase Server cluster
in combination with (bucket level) waitUntilReady, yours truly tried to be overly clever
and introduce a check which waits until the number of kv nodes in the config "nodesExt"
section equal the number of entries in "nodes". Most of the time this just works out
fine, but there are cases (both on-prem and in the cloud) where it doesn't.

Modifications
-------------
Since the check itself is flawed it needs to be removed in order to not time out on
perfectly healtyh on-prem and cloud clusters.

Change-Id: Ifcb455ef5c34115c883467ca2e42fc5e4ad848ae
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/177888
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>